LPD-94824 Move calendar view components and utils into subfolders
