+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/components/props_transformer/views/calendar_view/components/CalendarTaskCard.tsx
@@ -8,8 +8,8 @@ import {AssigneeAvatar} from '@liferay/object-dynamic-data-mapping-form-field-ty
 import classNames from 'classnames';
 import React from 'react';
 
-import isOverdue from '../../../../utils/isOverdue';
-import {ITaskObjectEntry} from '../../../../utils/types';
+import isOverdue from '../../../../../utils/isOverdue';
+import {ITaskObjectEntry} from '../../../../../utils/types';
 
 import './CalendarTaskCard.scss';
