+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/components/props_transformer/views/calendar_view/utils/unscheduledTasksAtom.ts
@@ -5,7 +5,7 @@
 
 import {State} from '@liferay/frontend-js-state-web';
 
-import {ITaskObjectEntry} from '../../../../utils/types';
+import {ITaskObjectEntry} from '../../../../../utils/types';
 
 export const CMP_UNSCHEDULED_TASKS_ATOM_ID = 'cmpUnscheduledTasks';
