+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/tests/js/CalendarMoreLinkPopover.spec.tsx
@@ -7,7 +7,7 @@ import '@testing-library/jest-dom';
 import {render} from '@testing-library/react';
 import React from 'react';
 
-import CalendarMoreLinkPopover from '../../js/components/props_transformer/views/calendar_view/CalendarMoreLinkPopover';
+import CalendarMoreLinkPopover from '../../js/components/props_transformer/views/calendar_view/components/CalendarMoreLinkPopover';
 import {ITaskObjectEntry} from '../../js/utils/types';
 
 jest.mock('@clayui/drop-down', () => ({
